chore: created different directory for each badge
