+++ src/components/badge/empty/index.tsx
@@ -1,0 +1,31 @@
+import { HTMLAttributes, forwardRef } from 'react'
+import {
+  BadgeBaseCVA,
+  EmptyBadgeSizeCVA,
+  BadgeEmptyClassProps,
+} from '../badge.classname'
+import qtMerge from 'qtMerge'
+
+export interface EmptyBadgeProps
+  extends Omit<HTMLAttributes<HTMLDivElement>, 'children'>,
+    BadgeEmptyClassProps {}
+
+export const EmptyBadge = forwardRef<HTMLDivElement, EmptyBadgeProps>(
+  ({ size, className, colorStyle, ...rest }, ref) => {
+    return (
+      <span
+        className={qtMerge(
+          BadgeBaseCVA({
+            colorStyle,
+          }),
+          EmptyBadgeSizeCVA({ size }),
+          className,
+        )}
+        ref={ref}
+        {...rest}
+      />
+    )
+  },
+)
+
+EmptyBadge.displayName = 'EmptyBadge'
